+++ src/components/molecules/RadioField/RadioField.tsx
@@ -6,6 +6,16 @@ import InputLabel from '../../atoms/InputLabel/InputLabel';
 import SizedContainer from '../../layout/SizedContainer/SizedContainer';
 import { FieldProps, InputStatus, InputProps } from '../../types';
 
+const getCheckedColor = ({ disabled, isValid }: Pick<InputProps, 'disabled' | 'isValid'>) => {
+  if (isValid) {
+    return colors.success;
+  }
+  if (disabled) {
+    return colors.grey;
+  }
+  return colors.brand;
+};
+
 const zoomIn = keyframes`
   from {
     transform: scale(1.9);
@@ -26,21 +36,26 @@ const FieldContainer = styled(SizedContainer)`
 
 const Label = styled(InputLabel)<InputStatus>`
   display: flex;
-  align-items: center;
   line-height: 1.4;
   font-weight: ${typography.weights.regular};
   font-size: ${typography.sizes.text.body};
   color: ${colors.greyDarkest};
+  padding: 14px 16px;
+  border: 1px solid ${getBorderColorByStatus};
+  transition-property: border, box-shadow;
+  transition: 0.2s ease-in-out;
+  border-radius: 8px;
   font-weight: 400;
   position: relative;
   margin-bottom: 0;
+  text-align: left;
   &:before {
     content: '';
+    flex-shrink: 0;
     background-color: ${colors.white};
     border-radius: 50%;
-    height: 24px;
-    width: 24px;
-    min-width: 24px;
+    height: 20px;
+    width: 20px;
     display: inline-block;
     margin-right: 8px;
     border: 1px ${getBorderColorByStatus} solid;
@@ -53,8 +68,10 @@ const Label = styled(InputLabel)<InputStatus>`
     position: absolute;
     border-radius: 50%;
     display: inline-block;
-    left: 6px;
-    top: 6px;
+    transition-property: background-color;
+    transition: 0.2s ease-in-out;
+    left: 21px;
+    top: 19px;
   }
   &:hover {
     cursor: pointer;
@@ -71,30 +88,42 @@ const Input = styled.input<InputStatus>`
   z-index: -1;
   position: absolute;
   &:hover + label {
+    border-color: ${colors.brand};
     &:before {
       border-color: ${colors.brand};
     }
   }
+  &:checked + label {
+    border-color: ${getCheckedColor};
+    &:before {
+      border-color: ${getCheckedColor};
+    }
+    &:after {
+      background-color: ${getCheckedColor};
+      height: 10px;
+      width: 10px;
+      animation: ${zoomIn} 200ms ease-in-out;
+    }
+  }
   &:focus + label {
+    border-color: ${colors.brand};
+    box-shadow: 0 0 4px 0 ${colors.brand};
     &:before {
       border-color: ${colors.brand};
       box-shadow: 0 0 4px 0 ${colors.brand};
     }
   }
-  &:checked + label {
-    &:before {
-      border-color: ${colors.brand};
-    }
+  &:focus:checked + label {
     &:after {
-      background-color: ${({ disabled }) => (disabled ? colors.grey : colors.brand)};
-      height: 12px;
-      width: 12px;
-      animation: ${zoomIn} 200ms ease-in-out;
+      background-color: ${colors.brand};
     }
   }
   &:disabled + label {
     cursor: not-allowed;
     color: ${colors.grey};
+  }
+  &:disabled:not(:checked) + label {
+    border-color: ${colors.greyLight};
     &:before {
       border-color: ${colors.greyLight};
     }